将助力和参与活动部分重构，使用redis

+++ src/main/java/com/md/luck/lottery/common/entity/respons/CastCulpChild.java
@@ -12,5 +12,5 @@ public class CastCulpChild extends CastCulp {
     /**
      * 排名
      */
-    private Integer rank;
+    private Long rank;
 }
